remove duplicated tests
