+++ tests/callbacks/test_early_stopping.py
@@ -130,3 +130,49 @@ def test_pickling(tmpdir):
     early_stopping_pickled = cloudpickle.dumps(early_stopping)
     early_stopping_loaded = cloudpickle.loads(early_stopping_pickled)
     assert vars(early_stopping) == vars(early_stopping_loaded)
+
+
+def test_early_stopping_no_val_step(tmpdir):
+    """Test that early stopping callback falls back to training metrics when no validation defined."""
+
+    class CurrentModel(EvalModelTemplate):
+        def training_step(self, *args, **kwargs):
+            output = super().training_step(*args, **kwargs)
+            output.update({'my_train_metric': output['loss']})  # could be anything else
+            return output
+
+    model = CurrentModel()
+    model.validation_step = None
+    model.val_dataloader = None
+
+    stopping = EarlyStopping(monitor='my_train_metric', min_delta=0.1)
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        early_stop_callback=stopping,
+        overfit_batches=0.20,
+        max_epochs=2,
+    )
+    result = trainer.fit(model)
+
+    assert result == 1, 'training failed to complete'
+    assert trainer.current_epoch < trainer.max_epochs
+
+
+def test_early_stopping_functionality(tmpdir):
+
+    class CurrentModel(EvalModelTemplate):
+        def validation_epoch_end(self, outputs):
+            losses = [8, 4, 2, 3, 4, 5, 8, 10]
+            val_loss = losses[self.current_epoch]
+            return {'val_loss': torch.tensor(val_loss)}
+
+    model = CurrentModel()
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        early_stop_callback=True,
+        overfit_batches=0.20,
+        max_epochs=20,
+    )
+    trainer.fit(model)
+    assert trainer.current_epoch == 5, 'early_stopping failed'
